requirements.txt
# Discord.py and related
discord.py==2.3.2
aiohttp==3.9.1
aiosqlite==0.19.0

# Excel handling
openpyxl==3.1.2

# File operations
aiofiles==23.2.1

# System monitoring
psutil==5.9.6

# Web server for keep-alive
flask==3.0.0

# Date/time utilities
python-dateutil==2.8.2

# Additional utilities
asyncio